#17: исправления, в соотствии замечаниям
