--- src/main/java/com/simbirsoft/projectManager/security/config/WebSecurityConfig.java
@@ -1,65 +1,0 @@
-package com.simbirsoft.projectManager.security.config;
-
-import com.simbirsoft.projectManager.security.service.UserAuthService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-//TODO Закомментировал до появления шаблонов
-@Configuration
-@EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
-
-  private UserAuthService userAuthService;
-
-  @Autowired
-  public void setUserAuthService(UserAuthService userAuthService) {
-    this.userAuthService = userAuthService;
-  }
-
-  @Bean
-  public PasswordEncoder getPasswordEncoder() {
-    return new BCryptPasswordEncoder(8);
-  }
-
-  @Override
-  protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-    auth.authenticationProvider(authenticationProvider());
-  }
-
-  @Override
-  protected void configure(HttpSecurity http) throws Exception {
-    http
-        .csrf().disable()
-        .authorizeRequests()
-        .antMatchers("/login").permitAll()
-        .antMatchers("/register").permitAll()
-        .antMatchers("/**").authenticated()
-        .and()
-        .formLogin()
-        .loginPage("/login")
-        .loginProcessingUrl("/authenticateTheUser")
-        .and()
-        .logout()
-        .logoutSuccessUrl("/login")
-        .permitAll();
-
-  }
-
-  @Bean
-  public DaoAuthenticationProvider authenticationProvider() {
-    DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
-    authenticationProvider.setPasswordEncoder(getPasswordEncoder());
-    authenticationProvider.setUserDetailsService(userAuthService);
-    return authenticationProvider;
-  }
-}
